restructured mailchimp files
