+++ src/examples/mailchimp/src/backend/entities/category.ts
