+++ src/examples/mailchimp/src/backend/entities/index.ts
@@ -1,0 +1,4 @@
+export * from './category';
+export * from './interest';
+export * from './list';
+export * from './member';
