+++ src/examples/mailchimp/src/backend/entities/interest.ts
