+++ src/examples/mailchimp/src/backend/entities/member.ts
