+++ src/examples/mailchimp/src/backend/schema/category/entity.ts
@@ -1,6 +1,6 @@
 import { Field, GraphQLEntity, ID, ObjectType, SummaryField } from '@exogee/graphweaver';
 
-import { CategoryDataEntity } from './data-entity';
+import { CategoryDataEntity } from '../../entities';
 
 @ObjectType()
 export class Category extends GraphQLEntity<CategoryDataEntity> {
