+++ src/examples/mailchimp/src/backend/schema/interest/entity.ts
@@ -1,6 +1,6 @@
 import { Field, ID, GraphQLEntity, ObjectType, SummaryField } from '@exogee/graphweaver';
 
-import { InterestDataEntity } from './data-entity';
+import { InterestDataEntity } from '../../entities';
 
 @ObjectType()
 export class Interest extends GraphQLEntity<InterestDataEntity> {
