+++ src/examples/mailchimp/src/backend/schema/list/entity.ts
@@ -1,6 +1,6 @@
 import { Field, GraphQLEntity, ID, ObjectType, SummaryField } from '@exogee/graphweaver';
 
-import { MailingListDataEntity } from './data-entity';
+import { MailingListDataEntity } from '../../entities';
 
 @ObjectType()
 export class MailingList extends GraphQLEntity<MailingListDataEntity> {
